test: split home address modal into 2 files... try and get tests to pass on CI
